Use Git Large File Storage (LFS) for all binaries
This commit changes the repo to use Git Large File Storage (LFS) for
storing binary files. Previously, these binaries were stored in tree.
This should be avoided as binary diffs are not possible and Git will
always store the entire binary file for all times. For this reasons,
the size of the repo will grow very quickly and whenever doing a
`git clone` the entire history including all versions of all binaries
is downloaded.

With Git LFS, these binaries are stored on a different remote server
and the repo just contains diff-able text pointers to the binaries.
By using the `git-lfs` command line tool, the whole process can be made
completely transparent to the user.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>