Rework how the simple capture script handles capture failures
With this commit, the simple capture script now prints a warning and
retries to do the capture in case of a capture failure. This way, the
capture only finishes after successfully capturing the specified amount of
traces. Reasons for capture failure are communication failures between
scope and target or if the signal is too strong.

Previously, a failed capture just led to fewer traces being
collected and the user did not get informed.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>