Rework behavior of simple capture script in case of ADC saturation
The ADC on the ChipWhisperer Lite has an output range of [-0.5, 0.5).
Values outside of this range cause the ADC to saturate. Previously, if a
trace contained a sample < -0.45, the corresponding trace did not get
recorded without informing the user.

With this commit, a safe output range is defined as [-0.48, 0.48]. All
traces are recorded but when plotting the traces at the end of the capture,
a warning is displayed in case any of the collected traces contains values
outside this range. The user is then suggested to reduce the scope gain.
This allows the capture script to remain simple and gives the user a
quick way for verifying the setup before starting a long-running capture.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>