Avoid frequent segment array resizing by using correct size
By default, the ChipWhisperer API starts every new trace segment (in the
project file, not related to the segment/batch captured in one shot)
with the size of 1 trace. Upon appending new traces to the project, the
segment is resized in steps of 25 traces. This results in frequent array
resizing and decreasing capture rate.

With this commit, the allocation size is frequently checked. If it's
lower than the known segment size, it's corrected directly to this
value. The next resize operation will allocate sufficient memory for
the whole segment array.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>