Avoid decreasing capture rate during long captures
By default, the ChipWhisperer performs a range check `_updateRanges()`
on all previously captured traces whenever adding a new trace. As a
result, the complexity of the `append()` function scales linearly with
the number of previously captured traces which significantly slows down
the capture rate.

With this commit, all except for the latest two trace storage segments
are disabled. The `_updateRange()` function then behaves like a sliding
window and only looks at the latest two instead of all segments. This
allows to maintain high capture rates even when capturing millions of
traces.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>