[tvla] Compute all rounds, all bytes, add new argument for selection
This commit modifies the histograms & t-test computation as well as
the plotting to run for all rounds and all bytes by default. Using new
command line arguments, the user can specify a single round and/or a
single byte if needed. Plots for all computed t-test results are saved
and the script now prints results in text format, highlighting in which
rounds and bytes, leakage above the threshold has been detected.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>