[tvla] Handle empty sets
In case either the fixed or random set is empty, the statistics can't
be computed. This can happen e.g. if few traces are used only, or if
using the hamming distance as sensitive variable and analyzing the
initial round. In the latter case, the hamming distance can only be
zero or non-zero if the corresponding key byte is zero or non-zero,
respectively. Thus, either the fixed or the random set is empty.

Previously we anyway computed the statistics producing meaningless
results. With this commit, we don't try to compute the statistics
anymore if either of the sets is empty and report this to the user.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>