Update binaries to lowRISC/OpenTitan@862155839
This commit updates all binaries to a more recent version of the OpenTitan
repository. All binaries are taken from lowRISC/OpenTitan@862155839
except for the CW305 bitstream for which also lowRISC/OpenTitan@b8f24c491
has been cherry picked to reduce the FPGA resource utilization.

The reason for doing this is that there have been major updates to the
hardware IPs and the software, as well as SPI that are incompatible
with the previous binaries included in this repo. This complicates
experiments with newer bitstreams or modified application binaries.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>

Update lowrisc_opentitan to lowRISC/opentitan@369cffc85

Update code from upstream repository
https://github.com/lowRISC/opentitan to revision
369cffc85db0e6d5a667676a6f89987b94210e70

* [util,sw] Reverse bootstrap hash word order (Miguel Osorio)

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>