[tvla] Optimize t-test computation
This commit switches to a t-test function provided by scipy which
operates on arrays. Also, the code around the t-test function is
refactored to make better use of numpy.ndarrays, to reduce the amount
of control code and to reduce the number of loops. All this allows for
a speedup of 20x for the t-test computation.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>