Fix ChipWhisperer version, download archive instead of git checkout
Recently, we observed issues when using the ChipWhisperer toolchain in
editable mode as the main OpenTitan repository uses a different, older
version. As we can't update the version used in OpenTitan due to
dependency issues (Python 3.7 is required for newer ChipWhisperer
versions but not mandatory for OpenTitan). Instead this commit will just
install the newer version for people actually needing capture support
(and having a newer Python version anyway).

Also, for now we fix the version to stabilize the capture environment.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>