Adjust capture window for SHA3
During the first and third activity blocks, KMAC absorbs the fixed
prefix and the uncritical message, respectively. For SCA we are mainly
interested in the middle activity block when KMAC XORs the key into
the state and then performs the key absorption. To optimize capture
and processing, this commit adjust the capture offset and number of
samples to only capture this middle activity block.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>