Switch to virtual Python env and install ChipWhisperer in editable mode
The officially recommended way of using ChipWhisperer is to install it
in editable mode. We haven't done that so far because for OpenTitan a
different version is installed (due to Python 3.6) and in non-editable
mode (to speed up CI).

To again use ChipWhisperer in editable mode for ot-sca as recommended
and to avoid issues due to having different versions installed in
non-editable and editable mode, this commit now switches to a virtual
Python environment for ot-sca.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>