[tvla] Optimize performance and memory consumption of specific AES TVLA
Previously, we were generating 9 histograms one for each possible
Hamming weight of the sensitive variable. But the actual statistics
are then computed by combining the 8 histograms for the non-zero Hamming
weights (fixed vs. random). It this therefore more efficient to only
generate 2 histograms in the first place. One for Hamming weight = 0 and
one for Hamming weight > 0. Most importantly, this leads to a reduction
in memory footprint of roughly 4.5x.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>