Add typer CLI and yaml file support to tvla script.
Default configurations are specified in the script. The options
may be changed with yaml file and it is possible to overwrite them
using command line interface. This is provided by typer library.
It also provides descriptions when typed --help. Updated
configurations may be written into the .yaml file after analysis.

Signed-off-by: Abdullah Varici <abdullah.varici@lowrisc.org>