Trigger lintpy workflow on pull requests
Without the `pull_request` event, the workflow would only trigger on
pushes to branches in the repository.  When pushing to a forked branch,
the workflow would not trigger.  Thus, the `pull_request` event is
required to trigger the workflow in PRs.  The `push` event restricted to
the `master` branch additionally covers pushes directly to the master
branch, without creating a duplicate check on pushes to pull request
branches.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>