Add test for general TVLA on given histograms
This adds two tests: each test takes a histogram as input, runs TVLA on
it, and checks whether the resulting t-test traces contains a
significant deviation from the mean, which would indicate leakage.  For
one histogram we expect leakage, for the other we do not.

These tests are automatically discovered by pytest.  Run `pytest test`
in the top-level directory to execute them.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>