Compress TVLA general test data
This commit compresses the existing test data for the general TVLA test
with `np.savez_compressed()`.  No changes are required to the code, which
can continue to load the data with a simple `np.load()`.  The total file
size of the existing tests is reduced from >150 MiB to <1 MiB.  The
execution time of the tests does not increase measurably.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>