[docker/podman] Switch from docker to podman
This commit does the initial switch from docker to podman to not require
root permissions on the host. At the moment, the container can still be
run using docker.

Some more work would be required to simplify the container setup. But this
will make the image incompatible with docker. We can do this at a later
point in time if we really want to.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>