[util/device.py] Wait for ADC lock instead of aborting after first trial
The scope uses the clock provided from the target and depending on the
target, it may take some more time for the ADC to actually lock on the
clock. With this commit, we wait for up to 1.5s for the ADC to lock
instead of just aborting the capture if the ADC hasn't locked yet when
first checked.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>