[capture] Change fixed plaintext in SHA3 fixed vs. random batch capture
The previous value specified for the data/plaintext was actually the
value for the key defined in the DTR for fixed vs. random key TVLA. It
was non-ideal because depending on place & route it could introduce a
considerable amount of fake leakage due to the power rails being less
discharged upon loading the fixed plaintext compared to upon loading a
random plaintext. This can result in a vertical offset between the trace
sets which may cover actual leakage happening due to the SHA3
processing.

For this reason, this commit changes the fixed plaintext value to the
value specified under Section 5.3 in the DTR for fixed vs. random data
TVLA. This value may also show similar behavior but not to the same
extent. This commit thus also adds hand-crafted value (in a comment)
which behaves better and can be used for debugging.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>