[sca/capture] Add ctrl-c handler to not lose already captured data
In case a huge number of traces is captured, this handler allows the
user, in case capturing takes too much time, to abort the capture without
losing previously recorded data.
To use this handler it is registered within each of the capture
functions, where a Chipwhisperer project is used to store traces.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>