[otbn/capture] Add fvsr key test for ECC256 keygen
This commit adds the option to capture datasets for fixed vs random
key TVLA analysis for ECC256 keygen algorithm on OTBN.
This option is hidden behing the test_type parameter.
Currently this option is only enabled for normal capture mode in
capture_otbn_vertical() function.

The method is described in more detail in the discussion document
https://docs.google.com/document/d/1A4l_Bnkivyh5aOjFHNmbR7_BgILdGnik30iNk4cGss4/

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>