[tvla] Set HAMMING_WEIGHT as the default leakage model
AES leakage can be computed using either HAMMING_WEIGHT or
HAMMING_DISTANCE. HAMMING_DISTANCE was useful in the early stages
of the project for analysing unprotected implemtation and debugging
the setup. However, now we use only HAMMING_WEIGHT model for TVLA
so it makes sense to make it the default.

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>