[tvla] Add compute_leakage_general
At the moment aes specific leakage is computed in the function
compute_leakage_aes(), while leakage for general TVLA is computed
directly in run_tvla().
This commit factors out this computation into a separate function.
This is also necessary because we plan to add more functions that
will need to call compute_leakage.

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>