[tvla] Change default value for general-tast
At the moment the default value of the `--general-test` parameter
is `True`. This argument cannot be changed to `False` via command
line, but only through the yaml configuration file.
Since we don't use configuaration files for TVLA in the CI, it is
currently impossible to add tests for aes-specific TVLA.
To enable adding these tests, this commit changes the default value
to `False` and adds `--general-test` to CI commands where needed.

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>