[tvla] Enable trace filtering for TVLA-general
Tvla has a mechanism for filtering out outlier traces from the
project. At the moment, this filtering is applied only to
aes specific TVLA. This commit enables filtering for general TVLA
as well.
As a result of applied filtering, the expected result of the CI
test test_general_aes_nonleaking_project() has changed. This
commit also updates the expected result of this test, in order
to leave the repository in a working state.

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>