[capture] Specify capture length and offsets in target clock cycles
This commit changes the capture setup to specify the horizontal
capture offset and the window length in target clock cycles rather than
samples. The corresponding values in terms of samples are computed
together with the sampling rate and oversampling ratio based on the
target's PLL frequency and internal clock generator settings.

For other scopes such as WaveRunner, the sampling rate needs to be
read out from the scope to compute these values. This is left as a TODO.

Co-authored-by: Vladimir Rozic <vrozic@lowrisc.org>
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>