[scope] Autoconvert cycles into samples
Previously, the conversion from cycles into samples happened in the
Husky scope driver. This PR moves this conversion into a library,
allowing the user to either specify offset and windowlength in
samples or cycles.
In addition, the sampling rate also can be changed in the config
file.

In the future, we want to read the sampling rate from the scope.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>