[fi] Reset OpenTitan on faulty behavior
Previously, we re-programmed the FPGA when a FI crashed OpenTitan
running on CW310. As this took quite some time, we now read the
INITB pin - this pin returns false or none when the FPGA is not
programmed or the bitstream CRC does not match.
In this case, we are re-programming the bistream.

When the bitstream is OK, we simply reset OpenTitan by using
the power on reset pin (POR_N is mapped to USB_A14 on CW310).

This significantly increases the FI rate.

Closes #249.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>