[tvla] Fix plotting issues with new database and metadata format
This commit fixes incompatibility of tvla plotting with the new database
and metadata format.
For this, additional metadata is stored in aes and otbn capture. Also, a
new wrapper function for plotting is added to the tvla script. Finally,
aes and otbn config files are adjusted.
Further, this commit adds code to also include metadata in plots of aes
specific test.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>