[tvla] Enable passing lists of selected rounds/bytes for specific AES
Previously, we could only:
- compute all rounds and all bytes
- select individual rounds but all bytes and vice versa
- select one round and one byte

This commit enables passing a list of bytes and rounds. All
combinations of the combined lists are then evaluated, e.g.
--round-select 0 --round-select 1 --byte-select 0 --byte-select 15
will evaluate Bytes 0 and 15 in both Round 0 and Round 1, i.e., there
will be for tests.

This is helps reducing the memory consumption and compute load if only
few rounds or bytes need to be analyzed as well as for CI.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>