[db] Fix bug in array slicing
Previously, retrieving parts of a dataset (e.g., plaintext, ciphertext, ...)
did not work as there was a bug in the array slicing logic.

Basically, if start and end or end was not provided, "None" is the default
value. However, the check in the if condition should be
`if (start is not None)` and not `if (start)`.

This PR fixes this issue.

Moreover, now it is also possible to select a single entry of a dataset,
e.g., `get_plaintexts(0)` returns the plaintext at DB ID 0.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>