[tvla] Optimize memory usage
This PR optimizes the memory consumption of the TVLA script by
freeing some memory objects. By setting an array to None, the
Python internal garbage collector takes care of actually freeing
the memory. With this PR, I've managed to reduce the mem.
consumption of certain parts of the script to -50%.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>