[analysis] Add trace preprocessing
This PR adds a trace preprocessing script allowing the user to
(a) filter traces using a max. amount of tolerable deviation
from the trace average and (b) align the traces using a
reference window. Currently, the script is tested with the
new trace database format. Moreover, I've decided to implement
this functionality outside of the TVLA script as this script is
already really overblown.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>