[capture] Add FPGA mode for AES
On the FPGA, the AES core automatically raises the SCA trigger we are
using for capturing traces. However, on the chip this functionality is
disabled.

This PR transmits the fpga_mode bit to the code running on the FPGA or
the chip. Based on this bit, the target code either manually raises the
trigger or lets the AES core handle automatically the trigger.

The target code is available in lowRISC/opentitan#21741 and the binary
sca_ujson_fpga_cw310 was created using lowRISC/opentitan@13da8f3 and
the command:
./bazelisk.sh build //sw/device/tests/crypto/cryptotest/firmware:firmware

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>