Add A1 bitstream and pentest firmware
This commit adds the OpenTitan Earl Grey A1 FPGA bitstream and the
pentest binary to the repository. The previous binaries for the Earl
Grey Z1 version are retained with the filename suffix z1.

The FPGA bitstream was generated from commit
https://github.com/lowRISC/opentitan/commit/282d863 and created with
the command:
  `/bazelisk.sh build //hw/bitstream/vivado:fpga_cw310_test_rom`

The pentest binary was generated from commit
https://github.com/lowRISC/opentitan/commit/b2cee40 and created with
the command:
  `./bazelisk.sh build //sw/device/tests/penetrationtests/firmware:firmware_fpga_cw310_test_rom`

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>
Co-authored-by: Pascal Nasahl <nasahlpa@lowrisc.org>